requirements.txt
fastapi==0.116.2
langchain_google_genai==2.1.12
python-dotenv==1.1.1
uvicorn==0.36.0
